+++ pkg/services/object/acl/v2/request.go
@@ -1,15 +1,10 @@
 package v2
 
 import (
-	"fmt"
-
-	icrypto "github.com/nspcc-dev/neofs-node/internal/crypto"
 	"github.com/nspcc-dev/neofs-sdk-go/bearer"
 	"github.com/nspcc-dev/neofs-sdk-go/container/acl"
 	cid "github.com/nspcc-dev/neofs-sdk-go/container/id"
 	oid "github.com/nspcc-dev/neofs-sdk-go/object/id"
-	protosession "github.com/nspcc-dev/neofs-sdk-go/proto/session"
-	sessionSDK "github.com/nspcc-dev/neofs-sdk-go/session"
 	"github.com/nspcc-dev/neofs-sdk-go/user"
 )
 
@@ -18,8 +13,7 @@ import (
 type RequestInfo struct {
 	basicACL    acl.Basic
 	requestRole acl.Role
-	operation   acl.Op  // put, get, head, etc.
-	cnrOwner    user.ID // container owner
+	operation   acl.Op // put, get, head, etc.
 
 	idCnr cid.ID
 
@@ -50,11 +44,6 @@ func (r *RequestInfo) SetSenderKey(senderKey []byte) {
 // Request returns raw API request.
 func (r RequestInfo) Request() any {
 	return r.srcRequest
-}
-
-// ContainerOwner returns owner if the container.
-func (r RequestInfo) ContainerOwner() user.ID {
-	return r.cnrOwner
 }
 
 // ObjectID return object ID.
@@ -100,32 +89,4 @@ func (r RequestInfo) Operation() acl.Op {
 // RequestRole returns request sender's role.
 func (r RequestInfo) RequestRole() acl.Role {
 	return r.requestRole
-}
-
-// MetaWithToken groups session and bearer tokens,
-// verification header and raw API request.
-type MetaWithToken struct {
-	vheader *protosession.RequestVerificationHeader
-	token   *sessionSDK.Object
-	bearer  *bearer.Token
-	src     any
-}
-
-// RequestOwner returns ownerID and its public key
-// according to internal meta information.
-func (r MetaWithToken) RequestOwner(fsChain icrypto.HistoricN3ScriptRunner) (*user.ID, []byte, error) {
-	// if session token is presented, use it as truth source
-	if r.token != nil {
-		// verify signature of session token
-		if err := icrypto.AuthenticateToken(r.token, fsChain); err != nil {
-			return nil, nil, fmt.Errorf("authenticate session token: %w", err)
-		}
-		tokenIssuer := r.token.Issuer()
-		return &tokenIssuer, r.token.IssuerPublicKeyBytes(), nil
-	}
-	idSender, key, err := icrypto.GetRequestAuthor(r.vheader)
-	if err != nil {
-		return nil, nil, err
-	}
-	return &idSender, key, nil
 }